fix admin photo upload to do proper error checking
the admin photo upload area wasn't doing proper error checking to make
sure the photo was created properly before attempting to create a new
choice for that photo.  In addition to now having more validation on the
choice side of things, we don't try to create a choice if the photo is
invalid.