add workaround for safari cookies in iframe
by default safari does not allow iframes to set cookies if the user did not
navigate to the domain that is serving the iframe source. This code attempts
to get around this limitation by checking to see if we can set cookies.  If
not, then we redirect to our domain and then back to original page.