upgrade to Bootstrap 2.3.2
Fixes an issues with the modal boxes on iOS devices.

The bootstrap-sass gem seemed to not work with this version of Rails. For
varying versions, it was throwing different errors when attempting to compile
the SCSS files. Instead, I've extracted the most recent SCSS files from the
gem and add them to our sass directory. Since the 2.3.x branch is unlikely to
receive more updates, we likely won't be upgrading any time soon.

[#66504606]