improve error messages after votes, skips, etc.
For any Ajax action that has an error we will parse the JSON to look for the
error name. Right now, we're looking only for CantFindSessionFromCookies since
we already had a message for that. If the user gets an error that we're not
explicitly looking for then it'll display a generic error message with that
error name. We also have error messages for timeouts of requests or requests
that got cancelled.

[#61768664]