update info parameter parsing to allow |
By default, URI.parse will fail if the URL includes a '|'. We can
customize the parser by making that an allowable character. The '|'
symbol is listed as "unsafe" by RFC 2396, but in this case shouldn't
cause problems. We had suggested using | as a delimiter when including
multiple data values via the "info" parameter. With this change, the
code now supports that recommendation.