add test to excercise some of CSV generation code
Excercising the full CSV generation code is a bit tricky with the
pairwise interaction. For now, we're just passing in dummy data to the
munge_csv_data method. Most of the code is getting executed, so we'd
catch things like syntax errors or non-existent method calls. We are not
validated the results of the method, though.